+++ scripts/run-tests.sh
@@ -7,11 +7,6 @@ source ./environment.test.sh
 
 export DATALOG_DIR="cbits/"
 
-make db-drop
-make db-setup
-
-cabal run -- flora-cli create-user --username "hackage-user" --email "tech@flora.pm" --password "foobar2000"
-
 if [ -z "$1" ] ;
 then
   cabal test